Aug 29, 2023, 7:27 AM

+++ ch_06/q14.java
@@ -10,7 +10,7 @@ public class q14 {
     }
     
     public static double estimatePI(int i) {
-        double m = (double) 4 * Math.pow(-1, i + 1) / (2 * i - 1);
+        double m = 4.0 * Math.pow(-1, i + 1) / (2.0 * i - 1);
         return m;
     }
 }
